[#782] moved the utility iOS only package
