+++ ios/CCC/Util/ObservableSEED.swift
@@ -46,12 +46,12 @@ final class ObservableSEED<
         logger.i(message: {"ObservableSEED \(ViewModel.description()) startObserving"})
 
         if viewModel.state != nil {
-            closeable = CoroutineUtilKt.observeWithCloseable(viewModel.state!, onChange: {
+            closeable = IOSCoroutineUtilKt.observeWithCloseable(viewModel.state!, onChange: {
                 self.state = $0 as! State
             })
         }
         if viewModel.effect != nil {
-            closeable = CoroutineUtilKt.observeWithCloseable(viewModel.effect!, onChange: {
+            closeable = IOSCoroutineUtilKt.observeWithCloseable(viewModel.effect!, onChange: {
                 self.effect.send($0 as! Effect)
             })
         }
